breaking out idiomatic computation graph detector and executor
